+++ src/formatters/trivia.rs
